Add files via upload
Single Gaussian

The outflow was separated into quadrants; primary and secondary cavity in the northern and southern region. The highest intensities and their x coordinates were recorded and a single gaussian function was fit to either side of the outflow. The data recorded into a text file was the location of the cut along the initial y axis, the location of the peak of the gaussian, its x coordinate and its sigma. The error of the peak, the location and the sigma were also found.

A linear fit was done to the intensity location and its x coordinate and these were also plotted on top of the original outflow.